Extract code out into separate files
This is pretty nice.  Note that src/main.rs is now 3 lines, and tests
are separated out into a bunch of different files.
